Complain when gcloud or ginkgo binaries are not found
Found when trying to run e2e with openstack. We seem to default
to "gce" as the provider, so it's ok to try to use gcloud, but
a better message would alert the user better that they are trying
to use "gce" which needs "gcloud". When ginkgo is missing, we should
prompt the user on the right way to get that binary.