Fix podRequest problem when init containers are present
Looks like an oversight introduced in 8b3bd5ae6033ffa7ff6b81b534031a6710bb8557
is causing a regression. When there are init containers, we should
loop through those and not the containers. also we are calculating the
wrong info since we are not updating initValue so we don't really
make sure podRequest takes into account "sum of init container
request(s)"